server refactored
