+++ packages/graphmarket-gateway/src/server/index.ts
@@ -1,4 +1,3 @@
-import listen from './listen';
-
+export { default } from './server';
+export { default as server } from './server';
 export { default as serviceList } from './serviceList';
-export default { listen };
