+++ packages/graphmarket-helpers/package.json
@@ -35,8 +35,7 @@
     "graphql-tag": "^2.11.0",
     "graphql-upload": "^11.0.0",
     "helmet": "^4.2.0",
-    "type-graphql": "^1.1.1",
-    "typedi": "^0.8.0"
+    "type-graphql": "^1.1.1"
   },
   "devDependencies": {
     "@types/compression": "^1.7.0",
