+++ packages/graphmarket-service-users/src/server/index.ts
@@ -1,5 +1,2 @@
-import connectDatabase from './connectDatabase';
-import initAdapters from './initAdapters';
-import listen from './listen';
-
-export default { connectDatabase, initAdapters, listen };
+export { default } from './server';
+export { default as server } from './server';
